+++ packages/anthropic/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/anthropic
 
+## 2.0.5
+
+### Patch Changes
+
+- Updated dependencies [68751f9]
+  - @ai-sdk/provider-utils@3.0.4
+
+## 2.0.4
+
+### Patch Changes
+
+- ae859ce: Added support for Anthropic provider's server-side code execution tool
+
 ## 2.0.3
 
 ### Patch Changes