+++ packages/openai-compatible/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ai-sdk/openai-compatible
 
+## 1.0.10
+
+### Patch Changes
+
+- 7ca3aee: refactoring(provider/openai-compatible): move models into folders
+
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [68751f9]
+  - @ai-sdk/provider-utils@3.0.4
+
+## 1.0.8
+
+### Patch Changes
+
+- 515c891: fix duplicated tool-input-start in certain cases
+
 ## 1.0.7
 
 ### Patch Changes