+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,34 @@
 
 ### Patch Changes
 
+- Updated dependencies [4176ecb]
+- Updated dependencies [20f23f9]
+  - ai@5.0.17
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@5.0.16
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [ca4f68f]
+  - ai@5.0.15
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [7729e32]
+  - ai@5.0.14
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [a7b2e66]
 - Updated dependencies [9bed210]
   - ai@5.0.13